Agregue componente de edicion de post de usuario

+++ frontend/src/components/Post/PostFormEditor.jsx
@@ -9,11 +9,23 @@ import MapComponent from "../MapComponent/MapComponent";
 
 export function PostFormEditor(props) {
   const { name, lastname, createdAt, _id } = useSelector((state) => state.user);
-  
+  const [selectedOption, setSelectedOption] = useState('');
+  const [selectedOptionLoc, setSelectedOptionLoc] = useState('');
   const params = useParams()
 
+  const initialState = {
+    name: "",
+    contact: "",
+    location: "",
+    petType: "",
+    description: "",
+    latitude: 0,
+    longitude: 0,
+  };
+
+  const [inputs, setInputs] = useState(initialState);
+
   const [datos, setDatos] = useState([])
-  const [loading, setLoading] = useState(true)
 
   
   function toBase64(arr) {
@@ -29,7 +41,8 @@ export function PostFormEditor(props) {
       try{
         const data = res.data
         console.log("loadPost res",data)
-        setDatos(data.publication)
+        const {name, img, contact, location,  petType, description,latitude, longitude} = res.data.publication;
+        setInputs({name, contact, location, petType, description,latitude, longitude});
         //proceso de imagen
         datos.map((i) => {
           const base64String = btoa(
@@ -37,42 +50,33 @@ export function PostFormEditor(props) {
               (data, byte) => data + String.fromCharCode(byte),
               ""
             )
-          ); 
-      
-          
-        }) 
+          );     
+        })
+        setImgPreview(`data:image/png;base64,${toBase64(data.publication.img.data.data)}`)
       } catch(error) {
         console.log(error)
       } 
     }
     loadPost()
   }, []);
-  console.log(" saliendo del loadPost Datos",datos)
   const [imgFile, setImgFile] = useState();
-  const [imgPreview, setImgPreview] = useState(iconoFoto);
+  const [imgPreview, setImgPreview] = useState();
 
   const navigate = useNavigate();
-
-
-  const initialState = {
-    name: "",
-    testImage: null,
-    contact: "",
-    location: "",
-    description: "",
-    user_id: _id,
-    nameUser: name,
-    lastnameUser: lastname,
-    createdAt: createdAt,
-    latitude: 0,
-    longitude: 0,
-  };
-
-  const [inputs, setInputs] = useState(initialState);
 
   const handleInputChange = (event) => {
     const { name, value } = event.target;
     setInputs({ ...inputs, [name]: value });
+  };
+
+  const handleSelectChange = (event) => {
+    setSelectedOption(event.target.value);
+    inputs.petType = event.target.value;
+  };
+
+  const handleSelectChangeLoc = (event) => {
+    setSelectedOptionLoc(event.target.value);
+    inputs.location = event.target.value;
   };
 
   const OnImgChange = (event) => {
@@ -95,21 +99,8 @@ export function PostFormEditor(props) {
 
   const handleSubmit = async (event) => {
     event.preventDefault();
-    const formData = new FormData();
-    formData.append("name", inputs.name);
-    formData.append("testImage", inputs.testImage);
-    formData.append("contact", inputs.contact);
-    formData.append("location", inputs.location);
-    formData.append("description", inputs.description);
-    formData.append("user_id", inputs.user_id);
-    formData.append("nameUser", inputs.nameUser);
-    formData.append("lastnameUser", inputs.lastnameUser);
-    formData.append("createdAt", inputs.createdAt);
-    formData.append("latitude", Center.lat);
-    formData.append("longitude", Center.lng);
-
     const updatePost = async () => {
-      const res = await servicePosts.updatePost(params.postId, formData);
+      const res = await servicePosts.updatePost(params.postId, inputs);
       try {
         console.log(res.data);
         navigate(`/userPosts`, { replace: true });
@@ -122,32 +113,35 @@ export function PostFormEditor(props) {
 
   const handleDeletePost = (event) =>{
     event.preventDefault();
-    const deletePost = async () =>{
-      const res = await servicePosts.deletePost(params.postId);
-      try {
-        console.log(res.data);
-        navigate(`/userPosts`, { replace: true });
-      } catch (error) {
-        console.log(error);
-      }
-    };
-    deletePost();
+    let isDelete = window.confirm(`¿de verdad deseas eliminar el post?`)
+    if(isDelete){
+      const deletePost = async () =>{
+        const res = await servicePosts.deletePost(params.postId);
+        try {
+          console.log(res.data);
+          navigate(`/userPosts`, { replace: true });
+        } catch (error) {
+          console.log(error);
+        }
+      };
+      deletePost();
+    }
   };
 
   
   const [ubicState, setUbicState] = useState(true);
   
   const [Center , setCenter] = useState({ 
-    lat: Number(datos.latitude),
-    lng: Number(datos.longitude),
+    lat: Number(inputs.latitude),
+    lng: Number(inputs.longitude),
   })
   
   const [Marcadores, setMarcadores] = useState([
     {
       id: 1,
       position: {
-        lat: Number(datos.latitude),
-        lng: Number(datos.longitude),
+        lat: Number(inputs.latitude),
+        lng: Number(inputs.longitude),
       },
     },
   ]);
@@ -190,7 +184,7 @@ export function PostFormEditor(props) {
     <div className="flex flex-col  sm:flex-row w-fit mx-auto">
       <div className="flex flex-col sm:w-1/2">
         <label className="mb-2 font-medium  text-2xl sm:text-3xl">
-          Nombre de la mascota {params.postId}
+          Nombre de la mascota
         </label>
         <div className="mb-2 w-fit border-b-4 border-b-yellow-200">
         <input
@@ -217,6 +211,24 @@ export function PostFormEditor(props) {
             onChange={handleInputChange}
           />  
         </div>
+
+        <label className="mb-2 font-medium text-2xl sm:text-3xl">
+            Tipo de Mascota:
+          </label>
+
+          <div className="mb-2 w-fit p-[5px]">
+            <select value={selectedOption} onChange={handleSelectChange}>
+              <option value={inputs.petType}>
+                {inputs.petType}
+              </option>
+              <option name="petType" value="perro">
+                perro
+              </option>
+              <option name="petType" value="gato">
+                gato
+              </option>
+            </select>
+          </div>
         
         <label className="mb-2 font-medium text-2xl sm:text-3xl">
           Descripcion detallada:
@@ -237,13 +249,6 @@ export function PostFormEditor(props) {
         </div>
 
         <div className="mb-2 w-[100vw] border-b-4 border-b-yellow-200 sm:w-[30vw]">
-          <input
-              className="w-0 h-0 cursor-pointer border-none "
-              type="file"
-              id="img"
-              name="testImage"
-              onChange={OnImgChange}
-          />
           <label className="mb-2 w-auto font-medium text-2xl sm:text-3xl" htmlFor="img">
             Foto de la mascota:
                   {/* hay un problema con el margin pq la imagen esta dentro del div y no esta bien centrada ni idea como resolver */}
@@ -251,25 +256,36 @@ export function PostFormEditor(props) {
                   {/* <button className="relative top-0 inset-x-auto w-10 h-10 bg-yellow-300 rounded-full text-2xl font-bold text-white group-hover:bg-yellow-400 transition duration-300 ease-in-out transform group-hover:-translate-y-1 group-hover:scale-110"/> */}
             
           </label>
+          <p>
+              Aviso: Para cambiar la foto debe eliminar el post y crear uno nuevo
+          </p>
           
           
         </div>
       </div>
       <div className="flex flex-col sm:w-1/2 w-fit">
-        <label className="mb-2 font-medium text-2xl sm:text-3xl">
-          Zona donde fue visto por ultima vez
-        </label>
-        
-        <div className="mb-2 w-fit border-b-4 border-b-yellow-200">
-          <input
-            className="mb-2 focus:border-yellow w-[50vw] bg-white-black rounded-md px-1 text-lg sm:w-[30vw]"
-            type="text"
-            id="location"
-            name="location"
-            value={inputs.location}
-            onChange={handleInputChange}
-          />
-        </div>
+          <label className="mb-2 font-medium text-2xl sm:text-3xl">
+            Localidad:
+          </label>
+          <div className="mb-2 w-fit p-[5px]">
+            <select value={selectedOptionLoc} onChange={handleSelectChangeLoc}>
+              <option value={inputs.location}>
+                {inputs.location}
+              </option>
+              <option
+                name="location"
+                value="Corrientes"
+              >
+                Corrientes
+              </option>
+              <option
+                name="location"
+                value="Resistencia"
+              >
+                Resistencia
+              </option>
+            </select>
+          </div>
 
         <label className="mb-2 font-medium text-md sm:text-lg">
           Arrastre en el mapa a la ubicacion donde fue visto por ultima vez:
